+++ apps/renderer/src/modules/feed-column/list.electron.tsx
@@ -1,0 +1,1 @@
+export { FeedList } from "./list.desktop"
